requirements.txt
flask
flask-cors
pandas
scikit-learn
joblib
numpy
openpyxl
python-dotenv